+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - Replaced `node-sass` with `sass` to allow using newer nodeJS versions.
 - Storybook is using webpack 5 now.
 - Exporting studyAPI and reudx actions so these can be used from extension components.
+- Survey component is using isLoggedIn attribute on survey contextya.
 
 ### Fixes